# meta/api/openapi.yaml
openapi: 3.1.0
info:
  title: Discobot Meta API
  version: 0.1.0
  summary: Central metadata, identity, authorization, and secret envelope API for Discobot.
  description: |
    Meta is Discobot's central metadata service. It owns organizations, users,
    devices, groups, projects, agent sessions, encrypted secret envelopes, secret
    bindings, and route-level authorization decisions.

    This specification starts with routes only. Request and response schemas are
    intentionally placeholders and should be replaced with shared component
    schemas once the route surface stabilizes.
servers:
  - url: /
    description: Current Meta service origin. Shortcut routes without `/v1/org/{organizationDomain}` resolve to the public organization.
...
paths:
...
  /authorize:
    get:
      tags: [OIDC]
      operationId: authorize
      x-discobot-action: oauth.authorize
      summary: Start or continue an OIDC authorization flow.
      description: |
        Browser-based OIDC authorization endpoint. The user authenticates to
        Meta or an upstream identity provider. On success, Meta redirects back to
        the client's redirect URI with an authorization code.
      parameters:
        - name: client_id
          in: query
          required: true
          schema: { type: string }
          description: OAuth client ID.
        - name: redirect_uri
          in: query
          required: true
          schema: { type: string, format: uri }
          description: Registered redirect URI.
        - name: response_type
          in: query
          required: true
          schema: { type: string, const: code }
          description: Must be `code` for authorization-code flow.
        - name: scope
          in: query
          required: true
          schema: { type: string }
          description: Space-delimited OAuth/OIDC scopes.
        - name: state
          in: query
          required: false
          schema: { type: string }
          description: Client-provided CSRF/state value returned with the redirect.
        - name: nonce
          in: query
          required: false
          schema: { type: string }
          description: OIDC nonce included in issued ID token.
        - name: code_challenge
          in: query
          required: false
          schema: { type: string }
          description: PKCE code challenge.
        - name: code_challenge_method
          in: query
          required: false
          schema: { type: string, enum: [S256] }
          description: PKCE code challenge method.
      responses:
        "302":
          description: Redirects to the client redirect URI or upstream identity provider.
        "400":
          description: Invalid authorization request.
  /token:
    post:
      tags: [OIDC, Tokens]
      operationId: token
      x-discobot-action: oauth.token
      summary: Exchange OAuth tokens.
      description: |
        OAuth token endpoint. Supports authorization-code exchange for normal
        login and OAuth token exchange for minting short-lived, agent-scoped
        access tokens.

        For user-to-agent communication, clients should request token exchange
        with an audience like `agent-session:{sessionId}`. Agents then accept
        only tokens whose audience matches their own agent session ID.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Token response containing ID/access/refresh tokens or exchanged access token.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid token request.
        "401":
          description: Invalid client or subject token.
  /userinfo:
    get:
      tags: [OIDC]
      operationId: getUserInfo
      x-discobot-action: userinfo.read
      summary: Get OIDC userinfo claims.
      description: |
        Returns user profile claims for the bearer access token. This endpoint is
        for OIDC client compatibility. Agent API authorization should use
        agent-scoped access tokens and local JWT validation instead.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Userinfo claims.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Missing or invalid bearer token.
  /logout:
    post:
      tags: [OIDC, Users]
      operationId: logout
      x-discobot-action: logout
      summary: Revoke the current user login session.
      description: |
        Revokes the current user session or performs provider logout behavior.
        Existing short-lived access tokens may remain valid until expiry unless
        agents introspect them.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Current login session revoked.
        "401":
          description: Missing or invalid bearer token.
...
  /v1/org:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      x-discobot-action: organization.list
      summary: List organizations visible to the caller.
      description: |
        Lists organizations the caller belongs to. The special organization with
        domain `public` is the default public organization used by shortcut routes
        such as `/v1/projects`.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Organizations.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      tags: [Organizations]
      operationId: createOrganization
      x-discobot-action: organization.create
      summary: Create an organization.
      description: Creates a top-level Meta organization.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Organization created.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /v1/org/{organizationDomain}:
    get:
      tags: [Organizations]
      operationId: getOrganization
      x-discobot-action: organization.read
      summary: Get organization metadata.
      description: Returns metadata for an organization. `organizationDomain` is the route-facing organization domain, such as `example.com`, or the special `public` organization.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
      responses:
        "200":
          description: Organization metadata.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      x-discobot-action: organization.update
      summary: Update organization metadata.
      description: Updates organization metadata according to organization admin rules.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Organization updated.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /v1/org/{organizationDomain}/members:
    get:
      tags: [Organizations]
      operationId: listOrganizationMembers
      x-discobot-action: organization.member.list
      summary: List organization members.
      description: Lists users that belong to an organization.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
      responses:
        "200":
          description: Organization members.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      tags: [Organizations]
      operationId: addOrganizationMember
      x-discobot-action: organization.member.add
      summary: Add an organization member.
      description: Adds a user to an organization with an organization role.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Organization member added.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /v1/org/{organizationDomain}/members/{userId}:
    patch:
      tags: [Organizations]
      operationId: updateOrganizationMember
      x-discobot-action: organization.member.update
      summary: Update organization member role.
      description: Updates an organization member role.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Organization member updated.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    delete:
      tags: [Organizations]
      operationId: removeOrganizationMember
      x-discobot-action: organization.member.remove
      summary: Remove organization member.
      description: Removes a user from an organization. Last-owner protections apply.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
        - $ref: '#/components/parameters/UserId'
      responses:
        "204":
          description: Organization member removed.
  /v1/org/{organizationDomain}/oauth-applications:
    get:
      tags: [Organizations]
      operationId: listOrganizationOAuthApplications
      x-discobot-action: oauth_application.list
      summary: List OAuth applications in an organization.
      description: Lists organization-scoped OAuth/OIDC client applications. Public organization admins use this to manage the default public login surface.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
      responses:
        "200":
          description: Organization OAuth applications.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      tags: [Organizations]
      operationId: createOrganizationOAuthApplication
      x-discobot-action: oauth_application.create
      x-discobot-apply:
        resource: OAuthApplication
        listOperationId: listOrganizationOAuthApplications
        getOperationId: getOrganizationOAuthApplication
        updateOperationId: updateOrganizationOAuthApplication
        scopeField: organization
        scopeParam: organizationDomain
        idField: id
        idParam: oauthApplicationId
        nameField: name
        matchFields: [provider]
      summary: Create an OAuth application in an organization.
      description: Creates an organization-scoped OAuth/OIDC client application.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthApplicationCreateRequest'
      responses:
        "201":
          description: OAuth application created.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /v1/org/{organizationDomain}/oauth-applications/{oauthApplicationId}:
    get:
      tags: [Organizations]
      operationId: getOrganizationOAuthApplication
      x-discobot-action: oauth_application.read
      summary: Get OAuth application metadata.
      description: Returns metadata for one organization-scoped OAuth/OIDC client application.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
        - $ref: '#/components/parameters/OAuthApplicationId'
      responses:
        "200":
          description: OAuth application metadata.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    patch:
      tags: [Organizations]
      operationId: updateOrganizationOAuthApplication
      x-discobot-action: oauth_application.update
      summary: Update OAuth application metadata.
      description: Updates redirect URIs, grant types, response types, scopes, auth method, name, or status for an organization-scoped OAuth application.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
        - $ref: '#/components/parameters/OAuthApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthApplicationUpdateRequest'
      responses:
        "200":
          description: OAuth application updated.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    delete:
      tags: [Organizations]
      operationId: deleteOrganizationOAuthApplication
      x-discobot-action: oauth_application.delete
      summary: Delete or disable an OAuth application.
      description: Deletes or disables an organization-scoped OAuth application.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationDomain'
        - $ref: '#/components/parameters/OAuthApplicationId'
      responses:
        "204":
          description: OAuth application deleted or disabled.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OAuthApplicationRequest:
      type: object
      properties:
        name:
          type: string
        provider:
          type: string
          enum: [github, google]
        clientId:
          type: string
        clientSecret:
          type: string
        redirectUris:
          type: array
          items:
            type: string
        grantTypes:
          type: array
          items:
            type: string
        responseTypes:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        tokenEndpointAuthMethod:
          type: string
        status:
          type: string
          enum: [active, disabled]
        providerConfig:
          type: object
          additionalProperties: true
        github:
          type: object
          additionalProperties: true
        google:
          type: object
          additionalProperties: true
    OAuthApplicationCreateRequest:
      allOf:
        - $ref: '#/components/schemas/OAuthApplicationRequest'
      required: [name, provider, clientId]
    OAuthApplicationUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/OAuthApplicationRequest'
  parameters:
...
    OAuthApplicationId:
      name: oauthApplicationId
      in: path
      required: true
      schema: { type: string }
      description: OAuth application ID.
    OrganizationDomain:
      name: organizationDomain
      in: path
      required: true
      schema: { type: string }
      description: Organization domain, such as `example.com`. The special `public` value addresses the public organization.
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
      description: User ID.
...